# website/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KnowledgeBank",
    "version": "v1"
  },
  "paths": {
...
    "/Roles/user/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Retrieve role of user",
        "description": "Retrieves the role of the given user ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The role of the user"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/User/all-users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets all users",
        "description": "Gets a list of all users.",
        "responses": {
          "200": {
            "description": "Users are loaded succesfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/User/list-paged": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Gets a page of users",
        "description": "Gets a page of users.",
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "searchQuery",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users are loaded succesfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/User/update-mail": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates the user email.",
        "description": "Updates the email of the user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User email updated successfully."
          },
          "400": {
            "description": "User email already exists."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/User/delete": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user.",
        "description": "Deletes a user by ID.",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UpdateEmailDto": {
        "required": [
          "email",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}